Fixed npe in holders, refactored to int location coords

+++ Client/src/main/java/arch/galaxyeclipse/client/ui/actor/ClickableActor.java
@@ -41,7 +41,7 @@ abstract class ClickableActor extends GeActor {
                     selected = !selected;
 
                 }
-                hitCommand.perform(new GePosition(x, y));
+                hitCommand.perform(new GePosition((int)x, (int)y));
             }
         });
